fix:matadata update youthnet l1 course pages
