+++ apps/learner-web-app/src/components/Content/L1ContentList.tsx
@@ -1,0 +1,284 @@
+'use client';
+import React from 'react';
+import Layout from '@learner/components/Layout';
+import LearnerCourse from '@learner/components/Content/LearnerCourse';
+import {
+  Box,
+  Button,
+  Grid,
+  Typography,
+  useMediaQuery,
+  useTheme,
+} from '@mui/material';
+import { gredientStyle } from '@learner/utils/style';
+import LTwoCourse from '@learner/components/Content/LTwoCourse';
+import { useEffect, useState } from 'react';
+import { getTenantInfo } from '@learner/utils/API/ProgramService';
+import ContentComponent from '@learner/components/Content/Content';
+import { useTranslation } from '@shared-lib';
+import { checkAuth } from '@shared-lib-v2/utils/AuthService';
+import { CompleteProfileBanner } from '@learner/components/CompleteProfileBanner/CompleteProfileBanner';
+import { profileComplitionCheck } from '@learner/utils/API/userService';
+import { usePathname } from 'next/navigation';
+import ArrowForwardIcon from '@mui/icons-material/ArrowForward';
+
+const MyComponent: React.FC = () => {
+  const pathname = usePathname();
+  const [filter, setFilter] = useState<Record<string, any> | null>(null);
+  const [isLogin, setIsLogin] = useState<boolean | null>(null);
+  const [isLoading, setIsLoading] = useState(true);
+  const [isProfileCard, setIsProfileCard] = useState(false);
+
+  useEffect(() => {
+    const fetchTenantInfo = async () => {
+      try {
+        if (checkAuth()) {
+          setIsLogin(true);
+          const result = await profileComplitionCheck();
+          console.log('Profile completion check result:', result);
+          setIsProfileCard(!result);
+        } else {
+          setIsLogin(false);
+        }
+        const res = await getTenantInfo();
+        const youthnetContentFilter = res?.result.find(
+          (program: any) => program.name === 'YouthNet'
+        );
+
+        const storedChannelId = localStorage.getItem('channelId');
+        if (!storedChannelId) {
+          const channelId = youthnetContentFilter?.channelId;
+          if (channelId) {
+            localStorage.setItem('channelId', channelId);
+          }
+        }
+
+        const storedTenantId = localStorage.getItem('tenantId');
+        if (!storedTenantId) {
+          const tenantId = youthnetContentFilter?.tenantId;
+          if (tenantId) {
+            localStorage.setItem('tenantId', tenantId);
+          }
+        }
+
+        const storedCollectionFramework = localStorage.getItem(
+          'collectionFramework'
+        );
+        if (!storedCollectionFramework) {
+          const collectionFramework =
+            youthnetContentFilter?.collectionFramework;
+          if (collectionFramework) {
+            localStorage.setItem('collectionFramework', collectionFramework);
+          }
+        }
+        setTimeout(() => {
+          setFilter({ filters: youthnetContentFilter?.contentFilter });
+          localStorage.setItem(
+            'filter',
+            JSON.stringify(youthnetContentFilter?.contentFilter)
+          );
+          setIsLoading(false);
+        }, 1000);
+      } catch (error) {
+        console.error('Failed to fetch tenant info:', error);
+      }
+    };
+    fetchTenantInfo();
+  }, [pathname]);
+
+  return (
+    <Layout isLoadingChildren={isLoading} sx={gredientStyle}>
+      {isProfileCard && <CompleteProfileBanner />}
+      {isLogin && (
+        <>
+          <Box
+            sx={{
+              height: 24,
+              display: 'flex',
+              alignItems: 'center',
+              py: '36px',
+              px: '34px',
+              bgcolor: '#fff',
+            }}
+          >
+            <Typography
+              variant="body1"
+              component="h2"
+              gutterBottom
+              sx={{
+                fontWeight: 500,
+                color: '#1F1B13',
+                textTransform: 'capitalize',
+              }}
+            >
+              <span role="img" aria-label="wave">
+                👋
+              </span>
+              Welcome, {localStorage.getItem('firstName')}!
+            </Typography>
+          </Box>
+          <InProgressContent />
+
+          <Grid container>
+            <Grid
+              item
+              xs={12}
+              md={12}
+              sx={{
+                display: 'flex',
+                flexDirection: 'column',
+                justifyContent: 'center',
+              }}
+            >
+              <LTwoCourse />
+            </Grid>
+          </Grid>
+        </>
+      )}
+
+      <Grid container style={gredientStyle}>
+        <Grid item xs={12}>
+          {filter && (
+            <LearnerCourse
+              title={'LEARNER_APP.COURSE.GET_STARTED'}
+              _content={{
+                pageName: 'L1_Content',
+                onlyFields: ['contentLanguage', 'se_subDomains', 'se_subjects'],
+                isOpenColapsed: [
+                  'contentLanguage',
+                  'se_subDomains',
+                  'se_subjects',
+                ],
+                staticFilter: {
+                  se_domains:
+                    typeof filter.filters?.domain === 'string'
+                      ? [filter.filters?.domain]
+                      : filter.filters?.domain,
+                  program:
+                    typeof filter.filters?.program === 'string'
+                      ? [filter.filters?.program]
+                      : filter.filters?.program,
+                },
+              }}
+            />
+          )}
+        </Grid>
+      </Grid>
+    </Layout>
+  );
+};
+
+export default MyComponent;
+
+const InProgressContent: React.FC = () => {
+  const { t } = useTranslation();
+  const [isShow, setIsShow] = useState(false);
+  const theme = useTheme();
+  // Detect if the screen size is medium or larger
+  const isMdUp = useMediaQuery(theme.breakpoints.up('md'));
+
+  return (
+    <Grid
+      container
+      style={gredientStyle}
+      sx={{ px: { xs: 0, sm: 0, md: 4 } }}
+      {...(isShow ? {} : { sx: { display: 'none' } })}
+    >
+      <Grid item xs={12} md={2.7}>
+        <Box
+          sx={{
+            display: 'flex',
+            flexDirection: { md: 'column' },
+            justifyContent: {
+              xs: 'space-between',
+              sm: 'space-between',
+              md: 'flex-start',
+            },
+            px: { xs: '16px' },
+            py: { xs: '24px' },
+            pb: { xs: 0, sm: 0 },
+            gap: 3,
+          }}
+        >
+          <Box
+            sx={{
+              display: 'flex',
+              flexDirection: 'column',
+              gap: { xs: 0, sm: 0, md: 1 },
+            }}
+          >
+            <Typography
+              variant="h4"
+              gutterBottom
+              sx={{
+                fontFamily: 'Poppins',
+                fontWeight: 400,
+                fontSize: { md: '22px', sm: '16px', xs: '16px' },
+                lineHeight: { md: '28px', sm: '24px', xs: '24px' },
+                letterSpacing: '0px',
+                verticalAlign: 'middle',
+                color: '#06A816',
+                mb: 0,
+              }}
+            >
+              {t('LEARNER_APP.L_ONE_COURSE.IN_PROGRESS_TITLE')}
+            </Typography>
+            <Typography
+              variant="body1"
+              gutterBottom
+              sx={{
+                color: '#7C766F',
+                fontSize: { xs: '12px', sm: '12px', md: '16px' },
+              }}
+            >
+              {t('LEARNER_APP.L_ONE_COURSE.ONGOING_COURSES').replace(
+                '{count}',
+                isShow?.toString()
+              )}
+            </Typography>
+          </Box>
+          <Box
+            sx={{
+              display: { md: 'flex' },
+              justifyContent: { md: 'flex-start' },
+            }}
+          >
+            <Button
+              variant={isMdUp ? 'contained' : 'text'}
+              sx={
+                !isMdUp
+                  ? {
+                      color: theme.palette.secondary.main,
+                      minWidth: '100px',
+                      fontWeight: 500,
+                      fontSize: '14px',
+                      lineHeight: '20px',
+                      letterSpacing: '0.1px',
+                      textAlign: 'center',
+                      verticalAlign: 'middle',
+                    }
+                  : {}
+              }
+              endIcon={<ArrowForwardIcon />}
+              color="primary"
+              href="/in-progress"
+            >
+              {t('LEARNER_APP.L_ONE_COURSE.VIEW_ALL_BUTTON')}
+            </Button>
+          </Box>
+        </Box>
+      </Grid>
+      <Grid item xs={12} md={9.3}>
+        <ContentComponent
+          getContentData={(e: any) => setIsShow(e.count)}
+          _config={{
+            isShowInCarousel: true,
+            isHideNavigation: true,
+            _subBox: { px: { xs: 2, sm: 2, md: 0 } },
+            _carousel: { spaceBetween: isMdUp ? 16 : 8 },
+          }}
+        />
+      </Grid>
+    </Grid>
+  );
+};
